huge refactor to better separate responsibilities
- UnityEditor configs now in Editor folder
- abstracted the config to separate responsibility of recording vs playback
- Controllers and Editors access user settings from the configs
