+++ Assets/Vimeo/Scripts/Config/PlayerSettings.cs
@@ -1,0 +1,7 @@
+namespace Vimeo.Config
+{
+    public class PlayerSettings : VimeoAuth
+    {
+        
+    }
+}
