+++ Assets/Vimeo/Scripts/Config/PlayerSettings.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 6c2bc3715e2d46e4f9642c6b45ec50bc
+timeCreated: 1518624887
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
