+++ Assets/Vimeo/Scripts/Config/RecorderSettings.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 26c1a3d082e4adf40a99de9453778154
-timeCreated: 1515451846
-licenseType: Pro
+guid: e932370820904ed49b5b2a05bd281c83
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
