+++ Assets/Vimeo/Scripts/Config/VimeoAuth.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d2833ade5c1f93942bb42bb4c9f72a42
+timeCreated: 1518624888
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
