+++ Assets/Vimeo/Scripts/Editor.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 8132a1182d8b78e42ae0c04ca7adcbea
+folderAsset: yes
+timeCreated: 1518562011
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
