+++ Assets/Vimeo/Scripts/Editor/BaseEditor.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 8a85f966c264cc240abdbe485566f289
+timeCreated: 1518624888
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
