+++ Assets/Vimeo/Scripts/Editor/RecorderEditor.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 63619801233ea42c295212c33e5fdfed
-timeCreated: 1508788897
-licenseType: Pro
+guid: a953a7c9ebd77164cb945b2335cac47c
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
