+++ Assets/Vimeo/Scripts/Editor/VimeoPlayerEditor.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: cec772ed6aa886b4ab79b18dfeee7df9
-timeCreated: 1508855978
+guid: de667601b96dfa9478e994723b2088ef
+timeCreated: 1518562011
 licenseType: Pro
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
