+++ Assets/Vimeo/Scripts/RecorderController.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: bb508fb5bcfd06a41baa0d8c8be842d6
+timeCreated: 1518624888
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
