+++ Assets/Vimeo/Scripts/RecorderInput.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace Vimeo
+{
+    
+}
