+++ Assets/Vimeo/Scripts/RecorderInput.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 035123c886acf784b934312f48b0b20e
+timeCreated: 1518562011
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
