+++ Assets/Vimeo/Scripts/Services/SlackApi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8910eac10d072c2428a2295aae6bcdda
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
