+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2017.3.1f1
+m_EditorVersion: 2018.1.0b7
